completing context reorganization. adding additional contexts for quadrat lists and census list
