+++ frontend/app/contexts/userselectioncontext.tsx
@@ -1,21 +1,41 @@
 "use client";
 import React, {createContext, Dispatch, useContext, useReducer} from "react";
-import {Plot, plots} from "@/config/macros";
+import {allCensus, allQuadrats, Plot, plots} from "@/config/macros";
 import {usePlotListContext} from "@/app/contexts/generalcontext";
 
 export const PlotsContext = createContext<Plot | null>(null);
+export const CensusContext = createContext<number | null>(null);
+export const QuadratContext = createContext<number | null>(null);
 export const PlotsDispatchContext = createContext<Dispatch<{ plotKey: string | null }> | null>(null);
+export const CensusDispatchContext = createContext<Dispatch<{ census: number | null }> | null>(null);
+export const QuadratDispatchContext = createContext<Dispatch<{ quadrat: number | null }> | null>(null);
 
 export default function PlotProvider({children}: { children: React.ReactNode }) {
   const [plot, plotDispatch] = useReducer(
     plotsReducer,
     null
   );
+  const [census, censusDispatch] = useReducer(
+    censusReducer,
+    null
+  );
+  const [quadrat, quadratDispatch] = useReducer(
+    quadratReducer,
+    null
+  );
   return (
     <>
       <PlotsContext.Provider value={plot}>
         <PlotsDispatchContext.Provider value={plotDispatch}>
-          {children}
+          <CensusContext.Provider value={census}>
+            <CensusDispatchContext.Provider value={censusDispatch}>
+              <QuadratContext.Provider value={quadrat}>
+                <QuadratDispatchContext.Provider value={quadratDispatch}>
+                  {children}
+                </QuadratDispatchContext.Provider>
+              </QuadratContext.Provider>
+            </CensusDispatchContext.Provider>
+          </CensusContext.Provider>
         </PlotsDispatchContext.Provider>
       </PlotsContext.Provider>
     </>
@@ -34,6 +54,17 @@ function plotsReducer(currentPlot: any, action: { plotKey: string | null }) {
     else return currentPlot;
   }
 }
+function censusReducer(currentCensus: any, action: { census: number | null }) {
+  if (action.census == null) return null;
+  else if (allCensus.includes(action.census)) return action.census;
+  else return currentCensus;
+}
+
+function quadratReducer(currentQuadrat: any, action: { quadrat: number | null }) {
+  if (action.quadrat == null) return null;
+  else if (allQuadrats.includes(action.quadrat)) return action.quadrat;
+  else return currentQuadrat;
+}
 
 export function usePlotContext() {
   return useContext(PlotsContext);
@@ -41,4 +72,19 @@ export function usePlotContext() {
 
 export function usePlotDispatch() {
   return useContext(PlotsDispatchContext);
+}
+export function useCensusContext() {
+  return useContext(CensusContext);
+}
+
+export function useCensusDispatch() {
+  return useContext(CensusDispatchContext);
+}
+
+export function useQuadratContext() {
+  return useContext(QuadratContext);
+}
+
+export function useQuadratDispatch() {
+  return useContext(QuadratDispatchContext);
 }
